Refactor `player` package -> `user`
