+++ bukkit/src/main/java/net/william278/husksync/BukkitHuskSync.java
@@ -41,9 +41,9 @@ import net.william278.husksync.listener.EventListener;
 import net.william278.husksync.migrator.LegacyMigrator;
 import net.william278.husksync.migrator.Migrator;
 import net.william278.husksync.migrator.MpdbMigrator;
-import net.william278.husksync.player.BukkitUser;
-import net.william278.husksync.player.ConsoleUser;
-import net.william278.husksync.player.OnlineUser;
+import net.william278.husksync.user.BukkitUser;
+import net.william278.husksync.user.ConsoleUser;
+import net.william278.husksync.user.OnlineUser;
 import net.william278.husksync.redis.RedisManager;
 import net.william278.husksync.util.BukkitLegacyConverter;
 import net.william278.husksync.util.BukkitTask;
