+++ bukkit/src/main/java/net/william278/husksync/event/BukkitEventDispatcher.java
@@ -20,8 +20,8 @@
 package net.william278.husksync.event;
 
 import net.william278.husksync.data.DataSnapshot;
-import net.william278.husksync.player.OnlineUser;
-import net.william278.husksync.player.User;
+import net.william278.husksync.user.OnlineUser;
+import net.william278.husksync.user.User;
 import org.bukkit.Bukkit;
 import org.jetbrains.annotations.NotNull;
