+++ bukkit/src/main/java/net/william278/husksync/event/BukkitPreSyncEvent.java
@@ -21,7 +21,7 @@ package net.william278.husksync.event;
 
 import net.william278.husksync.HuskSync;
 import net.william278.husksync.data.DataSnapshot;
-import net.william278.husksync.player.OnlineUser;
+import net.william278.husksync.user.OnlineUser;
 import org.bukkit.event.Cancellable;
 import org.bukkit.event.HandlerList;
 import org.jetbrains.annotations.NotNull;
