+++ bukkit/src/main/java/net/william278/husksync/listener/BukkitEventListener.java
@@ -22,8 +22,8 @@ package net.william278.husksync.listener;
 import net.william278.husksync.BukkitHuskSync;
 import net.william278.husksync.HuskSync;
 import net.william278.husksync.data.BukkitData;
-import net.william278.husksync.player.BukkitUser;
-import net.william278.husksync.player.OnlineUser;
+import net.william278.husksync.user.BukkitUser;
+import net.william278.husksync.user.OnlineUser;
 import org.bukkit.Bukkit;
 import org.bukkit.entity.Player;
 import org.bukkit.entity.Projectile;
