+++ bukkit/src/main/java/net/william278/husksync/migrator/LegacyMigrator.java
@@ -27,7 +27,7 @@ import net.william278.husksync.data.BukkitData;
 import net.william278.husksync.data.Data;
 import net.william278.husksync.data.DataSnapshot;
 import net.william278.husksync.data.Identifier;
-import net.william278.husksync.player.User;
+import net.william278.husksync.user.User;
 import net.william278.husksync.util.BukkitLegacyConverter;
 import org.bukkit.Material;
 import org.bukkit.Statistic;
