+++ common/src/main/java/net/william278/husksync/HuskSync.java
@@ -35,8 +35,8 @@ import net.william278.husksync.data.Serializer;
 import net.william278.husksync.database.Database;
 import net.william278.husksync.event.EventDispatcher;
 import net.william278.husksync.migrator.Migrator;
-import net.william278.husksync.player.ConsoleUser;
-import net.william278.husksync.player.OnlineUser;
+import net.william278.husksync.user.ConsoleUser;
+import net.william278.husksync.user.OnlineUser;
 import net.william278.husksync.redis.RedisManager;
 import net.william278.husksync.util.LegacyConverter;
 import net.william278.husksync.util.Task;
