+++ common/src/main/java/net/william278/husksync/command/Command.java
@@ -20,7 +20,7 @@
 package net.william278.husksync.command;
 
 import net.william278.husksync.HuskSync;
-import net.william278.husksync.player.CommandUser;
+import net.william278.husksync.user.CommandUser;
 import org.jetbrains.annotations.NotNull;
 
 import java.util.HashMap;
