+++ common/src/main/java/net/william278/husksync/command/InventoryCommand.java
@@ -23,8 +23,8 @@ import de.themoep.minedown.adventure.MineDown;
 import net.william278.husksync.HuskSync;
 import net.william278.husksync.data.Data;
 import net.william278.husksync.data.DataSnapshot;
-import net.william278.husksync.player.OnlineUser;
-import net.william278.husksync.player.User;
+import net.william278.husksync.user.OnlineUser;
+import net.william278.husksync.user.User;
 import org.jetbrains.annotations.NotNull;
 
 import java.time.format.DateTimeFormatter;
