+++ common/src/main/java/net/william278/husksync/command/ItemsCommand.java
@@ -21,9 +21,9 @@ package net.william278.husksync.command;
 
 import net.william278.husksync.HuskSync;
 import net.william278.husksync.data.DataSnapshot;
-import net.william278.husksync.player.CommandUser;
-import net.william278.husksync.player.OnlineUser;
-import net.william278.husksync.player.User;
+import net.william278.husksync.user.CommandUser;
+import net.william278.husksync.user.OnlineUser;
+import net.william278.husksync.user.User;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
