+++ common/src/main/java/net/william278/husksync/event/DataSaveEvent.java
@@ -21,7 +21,7 @@ package net.william278.husksync.event;
 
 import net.william278.husksync.HuskSync;
 import net.william278.husksync.data.DataSnapshot;
-import net.william278.husksync.player.User;
+import net.william278.husksync.user.User;
 import org.jetbrains.annotations.ApiStatus;
 import org.jetbrains.annotations.NotNull;
