+++ common/src/main/java/net/william278/husksync/user/ConsoleUser.java
@@ -17,48 +17,28 @@
  *  limitations under the License.
  */
 
-package net.william278.husksync.player;
+package net.william278.husksync.user;
 
+import net.kyori.adventure.audience.Audience;
 import org.jetbrains.annotations.NotNull;
 
-import java.util.UUID;
+public final class ConsoleUser implements CommandUser {
 
-/**
- * Represents a user who has their data synchronized by HuskSync
- */
-public class User {
-
-    private final UUID uuid;
-
-    private final String username;
-
-    public User(@NotNull UUID uuid, @NotNull String username) {
-        this.username = username;
-        this.uuid = uuid;
-    }
-
-    /**
-     * Get the user's unique account ID
-     */
     @NotNull
-    public UUID getUuid() {
-        return uuid;
-    }
+    private final Audience audience;
 
-    /**
-     * Get the user's username
-     */
-    @NotNull
-    public String getUsername() {
-        return username;
+    public ConsoleUser(@NotNull Audience console) {
+        this.audience = console;
     }
 
     @Override
-    public boolean equals(Object object) {
-        if (object instanceof User other) {
-            return this.getUuid().equals(other.getUuid());
-        }
-        return super.equals(object);
+    @NotNull
+    public Audience getAudience() {
+        return audience;
     }
 
+    @Override
+    public boolean hasPermission(@NotNull String permission) {
+        return true;
+    }
 }
