+++ common/src/main/java/net/william278/husksync/user/User.java
@@ -17,28 +17,48 @@
  *  limitations under the License.
  */
 
-package net.william278.husksync.player;
+package net.william278.husksync.user;
 
-import net.kyori.adventure.audience.Audience;
 import org.jetbrains.annotations.NotNull;
 
-public final class ConsoleUser implements CommandUser {
+import java.util.UUID;
 
+/**
+ * Represents a user who has their data synchronized by HuskSync
+ */
+public class User {
+
+    private final UUID uuid;
+
+    private final String username;
+
+    public User(@NotNull UUID uuid, @NotNull String username) {
+        this.username = username;
+        this.uuid = uuid;
+    }
+
+    /**
+     * Get the user's unique account ID
+     */
     @NotNull
-    private final Audience audience;
+    public UUID getUuid() {
+        return uuid;
+    }
 
-    public ConsoleUser(@NotNull Audience console) {
-        this.audience = console;
+    /**
+     * Get the user's username
+     */
+    @NotNull
+    public String getUsername() {
+        return username;
     }
 
     @Override
-    @NotNull
-    public Audience getAudience() {
-        return audience;
+    public boolean equals(Object object) {
+        if (object instanceof User other) {
+            return this.getUuid().equals(other.getUuid());
+        }
+        return super.equals(object);
     }
 
-    @Override
-    public boolean hasPermission(@NotNull String permission) {
-        return true;
-    }
 }
