+++ common/src/main/java/net/william278/husksync/util/DataDumper.java
@@ -23,7 +23,7 @@ import com.google.gson.JsonObject;
 import com.google.gson.JsonParser;
 import net.william278.husksync.HuskSync;
 import net.william278.husksync.data.DataSnapshot;
-import net.william278.husksync.player.User;
+import net.william278.husksync.user.User;
 import org.jetbrains.annotations.NotNull;
 
 import java.io.*;
